move to the separate module, rename metamethod
